move dto to core project
