--- Studenda.Core.Server/Security/Data/Transfer/LoginRequest.cs
@@ -1,7 +1,0 @@
-﻿namespace Studenda.Core.Server.Security.Data.Transfer;
-
-public class LoginRequest
-{
-    public string Email { get; init; } = null!;
-    public string Password { get; init; } = null!;
-}
