--- Studenda.Core.Server/Security/Data/Transfer/RegisterRequest.cs
@@ -1,8 +1,0 @@
-﻿namespace Studenda.Core.Server.Security.Data.Transfer;
-
-public class RegisterRequest
-{
-    public string Email { get; set; } = null!;
-
-    public string Password { get; set; } = null!;
-}
