+++ Studenda.Core/Data/Transfer/Security/LoginRequest.cs
@@ -1,0 +1,9 @@
+﻿namespace Studenda.Core.Data.Transfer.Security;
+
+/// <summary>
+///     Тело запроса авторизации.
+/// </summary>
+public class LoginRequest : SecurityRequest
+{
+    // PASS.
+}
