+++ Studenda.Core/Data/Transfer/Security/RegisterRequest.cs
@@ -1,0 +1,9 @@
+﻿namespace Studenda.Core.Data.Transfer.Security;
+
+/// <summary>
+///     Тело запроса регистрации.
+/// </summary>
+public class RegisterRequest : SecurityRequest
+{
+    // PASS.
+}
