+++ Studenda.Core/Studenda.Core.csproj
@@ -10,6 +10,4 @@
         <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="7.0.9"/>
     </ItemGroup>
 
-
-
 </Project>
